pyproject.toml
[tool.poetry]
name = "sw-core"
version = "0.1.0"
description = ""
authors = ["Slyusar Roman"]
readme = "README.md"
packages = [{ include = "sw_core" }]

[tool.poetry.dependencies]
python = "3.8.10"
sqlalchemy = "^2.0.17"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
tcp-latency = "^0.0.12"
schedule = "^1.1.0"
psycopg2 = "^2.9.6"
redmail = "^0.6.0"
aiofiles = "^23.1.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
